Moves the datasource/delegate into own files
Separates the UICollectionView and UITableView Delegate/DataSource objects into their own objects. This removes the need for FunctionalTableData to extend NSObject, isolates the logic/code associated with the protocols.
